+++ R/lavaan_cor.R
@@ -1,21 +1,18 @@
-#' @title Extract relevant covariance indices from lavaan model
+#' @title Extract relevant correlation indices from lavaan model
 #'
-#' @description Extract relevant covariance indices from lavaan model through
+#' @description Extract relevant correlation indices from lavaan model through
 #'              `lavaan::parameterEstimates` with `standardized = TRUE`. In this
-#'              case, the covariances represent the resulting
+#'              case, the correlation coefficient (r) represents the resulting
 #'              `std.all` column.
 #'
-#' @param fit lavaan fit object to extract covariance indices from
-#' @param estimate What estimate to use, either the standardized
-#'                 estimate ("B", default), or unstandardized
-#'                 estimate ("b").
+#' @param fit lavaan fit object to extract correlations from
 #' @param nice_table Logical, whether to print the table as a
 #'                   `rempsyc::nice_table` as well as print the
 #'                   reference values at the bottom of the table.
 #' @param ... Arguments to be passed to `rempsyc::nice_table`
 #' @keywords lavaan structural equation modeling path analysis CFA
-#' @return A dataframe of covariances, including the covaried variables,
-#'         the covariance, and corresponding p-value.
+#' @return A dataframe of correlations, including the correlated
+#'         variables, the correlation, and the corresponding p-value.
 #' @export
 #' @examplesIf requireNamespace("lavaan", quietly = TRUE)
 #' (latent <- list(visual = paste0("x", 1:3),
@@ -33,23 +30,15 @@
 #'
 #' library(lavaan)
 #' fit <- sem(HS.model, data=HolzingerSwineford1939)
-#' lavaan_cov(fit)
+#' lavaan_cor(fit)
 
-lavaan_cov <- function(fit, estimate = "B", nice_table = FALSE, ...) {
-  og.names <- c("lhs", "rhs", "pvalue", "est", "ci.lower", "ci.upper")
-  new.names <- c("Variable 1", "Variable 2", "p", "b", "CI_lower", "CI_upper")
-  if (estimate == "b") {
-    x <- lavaan::parameterEstimates(fit)
-  } else if (estimate == "B") {
-    x <- standardizedsolution(fit, level = 0.95)
-    og.names[4] <- "est.std"
-    new.names[4] <- "B"
-  } else {
-    stop("The 'estimate' argument may only be one of c('B', 'b').")
-  }
+lavaan_cor <- function(fit, nice_table = FALSE, ...) {
+  x <- lavaan::parameterEstimates(fit, standardized = TRUE)
   x <- x[which(x["op"] == "~~"),]
-  x <- x[og.names]
-  names(x) <- new.names
+  x <- x[c("lhs", "rhs", "std.all", "pvalue")]
+  not.cor <- which(x$lhs == x$rhs)
+  x <- x[-not.cor,]
+  names(x) <- c("Variable 1", "Variable 2", "r", "p")
   if (nice_table) {
     insight::check_if_installed("rempsyc", reason = "for this feature.")
     x <- rempsyc::nice_table(x, ...)