+++ R/nice_lavaanPlot.R
@@ -5,10 +5,12 @@
 #' @param model SEM or CFA model to plot.
 #' @param node_options Shape and font name.
 #' @param edge_options Colour of edges.
-#' @param coefs Logical, whether to plot coefficients.
-#' @param stand Logical, whether to standardized coefficients.
-#' @param covs Logical, whether to plot covariances
-#' @param stars Logical, whether to plot significance stars.
+#' @param coefs Logical, whether to plot coefficients. Defaults to TRUE.
+#' @param stand Logical, whether to use standardized coefficients.
+#'              Defaults to TRUE.
+#' @param covs Logical, whether to plot covariances. Defaults to FALSE.
+#' @param stars Which links to plot significance stars for. One of
+#'              `c("regress", "latent", "covs")`.
 #' @param sig Significance threshold.
 #' @param graph_options Read from left to right, rather than from top to bottom.
 #' @param ... Arguments to be passed to function `lavaanPlot::lavaanPlot`.
@@ -16,7 +18,7 @@
 #' @return A lavaanPlot, of classes `c("grViz", "htmlwidget")`, representing the
 #'         specified `lavaan` model.
 #' @export
-#' @examples
+#' @examplesIf requireNamespace("lavaan", quietly = TRUE) && requireNamespace("lavaanPlot", quietly = TRUE)
 #' (latent <- list(visual = paste0("x", 1:3),
 #'                 textual = paste0("x", 4:6),
 #'                 speed = paste0("x", 7:9)))
@@ -33,10 +35,10 @@
 
 nice_lavaanPlot <- function(
     model, node_options = list(shape = "box", fontname = "Helvetica"),
-    edge_options = list(color = "grey"), coefs = TRUE, stand = TRUE,
-    covs = FALSE, stars = list("regress"), sig = .05,
-    graph_options = list(rankdir = "LR"), ...) {
-  rlang::check_installed(c(
+    edge_options = c(color = "black"), coefs = TRUE, stand = TRUE,
+    covs = FALSE, stars = c("regress", "latent", "covs"), sig = .05,
+    graph_options = c(rankdir = "LR"), ...) {
+  insight::check_if_installed(c(
     "lavaanPlot", "DiagrammeRsvg", "rsvg", "png", "webshot"),
     reason = "for this function.")
   lavaanPlot::lavaanPlot(model = model,