+++ tests/testthat/test-nice_lavaanPlot.R
@@ -1,4 +1,4 @@
-library(lavaan)
+suppressWarnings(suppressPackageStartupMessages(library(lavaan)))
 
 # Define our other variables
 M <- "visual"
@@ -36,6 +36,7 @@ fit.sem2 <- sem(HS.model2, data)
 #   Tests                                                                   ####
 
 test_that("nice_lavaanPlot on CFA", {
+  skip_if_not_installed("lavaanPlot")
   expect_s3_class(
     nice_lavaanPlot(fit.cfa),
     c("grViz", "htmlwidget")
@@ -43,6 +44,7 @@ test_that("nice_lavaanPlot on CFA", {
 })
 
 test_that("nice_lavaanPlot on SEM", {
+  skip_if_not_installed("lavaanPlot")
   expect_s3_class(
     nice_lavaanPlot(fit.sem),
     c("grViz", "htmlwidget")
@@ -50,6 +52,7 @@ test_that("nice_lavaanPlot on SEM", {
 })
 
 test_that("nice_lavaanPlot on lavaan", {
+  skip_if_not_installed("lavaanPlot")
   expect_s3_class(
     nice_lavaanPlot(fit.lavaan),
     c("grViz", "htmlwidget")
@@ -57,6 +60,7 @@ test_that("nice_lavaanPlot on lavaan", {
 })
 
 test_that("nice_lavaanPlot different sem model", {
+  skip_if_not_installed("lavaanPlot")
   expect_s3_class(
     nice_lavaanPlot(fit.sem2),
     c("grViz", "htmlwidget")